pyproject.toml
[project]
name = "shioaji-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.2",
    "shioaji>=1.3.2",
]
